# Fix patching virtualOutbound-blackhole (#36329)