@istio-testing 2021-12-06
@howardjohn: #36366 failed to apply on top of branch "release-1.11":
```
Applying: Fix patching virtualOutbound-blackhole (#36329)
Using index info to reconstruct a base tree...
M	pilot/pkg/networking/core/v1alpha3/accesslog_test.go
M	pilot/pkg/networking/core/v1alpha3/listener_builder.go
M	pilot/test/xdstest/extract.go
Falling back to patching base and 3-way merge...
Auto-merging pilot/test/xdstest/extract.go
Auto-merging pilot/pkg/networking/core/v1alpha3/listener_builder.go
CONFLICT (content): Merge conflict in pilot/pkg/networking/core/v1alpha3/listener_builder.go
Auto-merging pilot/pkg/networking/core/v1alpha3/accesslog_test.go
error: Failed to merge in the changes.
hint: Use 'git am --show-current-patch=diff' to see the failed patch
Patch failed at 0001 Fix patching virtualOutbound-blackhole (#36329)
When you have resolved this problem, run "git am --continue".
If you prefer to skip this patch, run "git am --skip" instead.
To restore the original branch and stop patching, run "git am --abort".

```

<details>

In response to [this](https://github.com/istio/istio/pull/36366#issuecomment-987102061):

>/retest
>/cherrypick release-1.11


Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository.
</details>