@istio-testing 2021-12-06
@howardjohn: new issue created for failed cherrypick: #36391

<details>

In response to [this](https://github.com/istio/istio/pull/36366#issuecomment-987102061):

>/retest
>/cherrypick release-1.11


Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository.
</details>